pyproject.toml
[project]
name = "glyph"
version = "0.1.0"
description = "Tile map viewer with download capabilities"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "albumentations",
    "lightning",
    "wandb",
    "Flask-CORS==4.0.0",
    "Pillow",
    "Flask",
    "torchvision",
    "transformers",
    "geopandas",
    "shapely",
    "tqdm",
    "requests",
    "python-dotenv",
    "gdown",
    "opencv-python",
    "lxml",
    "scikit-image",
    "simplification",
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "json_repair"
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
tile_download = "tile_download.download.fetch_tiles:console_main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "packages" }

[tool.setuptools.packages.find]
where = ["packages"]